workers_table/requirements.txt
beautifulsoup4
djangorestframework
